requirements.txt
# Core Streamlit - REQUERIDO
streamlit==1.47.1

# Data processing - REQUERIDO
pandas==2.3.1
numpy==2.3.2

# Document processing - REQUERIDO para procesar PDF y DOCX
python-docx==1.1.0
PyPDF2==3.0.1
docx2txt==0.8

# Environment and configuration - REQUERIDO
python-dotenv==1.0.0

# AI/Google Gemini - REQUERIDO para funcionalidad de IA
google-generativeai>=0.6.0

# NOTA: Versiones compatibles con Python 3.13 y ya instaladas 